Remove extraneous development note comments Move a function from nocgo to main file for disk package
